+++ src/api/noteapi.h
@@ -12,6 +12,7 @@ class NoteApi : public QObject, public Note {
     Q_PROPERTY(int id READ getId())
     Q_PROPERTY(QString name READ getName() WRITE setName)
     Q_PROPERTY(QString fileName READ getFileName())
+    Q_PROPERTY(QString fullNoteFilePath READ fullNoteFilePath())
     Q_PROPERTY(QString noteText READ getNoteText() WRITE setNoteText())
     Q_PROPERTY(QString decryptedNoteText READ getDecryptedNoteText()
     WRITE setDecryptedNoteText())