+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # QOwnNotes Changelog
 
 ## 16.08.16
+- you are now able to define custom actions that will be shown as menu 
+  entry in the scripting menu and optionally in the toolbar as button
+  (for [Issue #297](https://github.com/pbek/QOwnNotes/issues/297))
+    - take a look at the example at
+      [custom-actions.qml](https://github.com/pbek/QOwnNotes/blob/develop/doc/scripting/custom-actions.qml)
+    - there is a small documentation about the `registerCustomAction`
+      function at
+      [README.md](https://github.com/pbek/QOwnNotes/blob/develop/doc/scripting/README.md)
 - the warning about tags when copying or moving notes to a note
   subfolder will now be only displayed if one of the selected notes has
   a tag assigned (for [Issue #289](https://github.com/pbek/QOwnNotes/issues/289))